+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ This document summarizes the most important changes of the individual releases o
 * Bump various tool versions.
 * Include an example for `cace`.
 * Add `pyuvm`.
+* Adding BSIMCMG model for `ngspice`.
 * Remove `gdstk` due to build issues.
 
 ## 2024.05